ecg/requirements.txt
pyserial>=3.5
matplotlib>=3.5.0
numpy>=1.21.0
scipy>=1.7.0
pandas>=1.3.0 